WikiWhat/requirements.txt
pandas
requests
wikipedia-api
nltk
tiktoken
openai
pinecone-client
tqdm
chromadb
claude-api
nest-asyncio